DataObjects/Core_domain.php

+++ DataObjects/Core_domain.php
@@ -67,8 +67,10 @@ class Pman_Core_DataObjects_Core_domain extends DB_DataObject
         $cd->whereAddIn('id', $ids, 'int');
         $domains = $cd->fetchAll();
 
-        foreach($domains as $domain) {
-            $domain->beforeDelete(array(), $roo);
+        if(method_exists($this, 'beforeDelete')) {
+            foreach($domains as $domain) {
+                $domain->beforeDelete(array(), $roo);
+            }
         }
 
         // delete domains
@@ -83,8 +85,10 @@ class Pman_Core_DataObjects_Core_domain extends DB_DataObject
         die('test');
         $cd->query($sql);
 
-        foreach($domains as $domain) {
-            $domain->onDelete(array(), $roo);
+        if(method_exists($this, 'onDelete')) {
+            foreach($domains as $domain) {
+                $domain->onDelete(array(), $roo);
+            }
         }
     }
